// src/application/interface/port/prompt-style-port.interface.ts
export interface IPromptStylePortInterface {
	ACTIVE_POINTER_ASCII: string;
	ACTIVE_POINTER_UNICODE: string;
	COMPLETED_MARK_ASCII: string;
	COMPLETED_MARK_UNICODE: string;
	DESCRIPTION_INDENT_PREFIX: string;
	DISABLED_LABEL_SUFFIX: string;
	ERROR_PREFIX: string;
	FAILED_MARK_ASCII: string;
	FAILED_MARK_UNICODE: string;
	GROUP_INDENT_PREFIX: string;
	HINT_PREFIX: string;
	IS_BLANK_LINE_AFTER_QUESTION_ENABLED: boolean;
	NEUTRAL_MARK_ASCII: string;
	NEUTRAL_MARK_UNICODE: string;
	PLACEHOLDER_PREFIX: string;
	QUESTION_PREFIX_SYMBOL: string;
	SELECTED_MARK_ASCII: string;
	SELECTED_MARK_UNICODE: string;
	SEPARATOR_SYMBOL: string;
	TOGGLE_FALSE_LABEL: string;
	TOGGLE_TRUE_LABEL: string;
	UNSELECTED_MARK_ASCII: string;
	UNSELECTED_MARK_UNICODE: string;
}